Cleanup prop tests
