+++ external-crates/move/crates/bytecode-verifier-tests/src/unit_tests/signature_tests.rs
@@ -26,39 +26,6 @@ fn test_reference_of_reference() {
     assert!(errors.is_err());
 }
 
-proptest! {
-    #[test]
-    fn valid_signatures(module in CompiledModule::valid_strategy(20)) {
-        prop_assert!(SignatureChecker::verify_module(&module).is_ok())
-    }
-
-    #[test]
-    fn double_refs(
-        mut module in CompiledModule::valid_strategy(20),
-        mutations in vec((any::<PropIndex>(), any::<PropIndex>()), 0..20),
-    ) {
-        let context = SignatureRefMutation::new(&mut module, mutations);
-        let expected_violations = context.apply();
-
-        let result = SignatureChecker::verify_module(&module);
-
-        prop_assert_eq!(expected_violations, result.is_err());
-    }
-
-    #[test]
-    fn field_def_references(
-        mut module in CompiledModule::valid_strategy(20),
-        mutations in vec((any::<PropIndex>(), any::<PropIndex>()), 0..40),
-    ) {
-        let context = FieldRefMutation::new(&mut module, mutations);
-        let expected_violations = context.apply();
-
-        let result = SignatureChecker::verify_module(&module);
-
-        prop_assert_eq!(expected_violations, result.is_err());
-    }
-}
-
 #[test]
 fn no_verify_locals_good() {
     let compiled_module_good = CompiledModule {
